Decouple black scholes

+++ contracts/Option/IOddzAsset.sol
@@ -17,7 +17,7 @@ interface IOddzAsset {
         bytes32 _name;
         address _address;
         bool _active;
-        uint256 _precision;
+        uint8 _precision;
     }
 
     struct AssetPair {
@@ -37,7 +37,7 @@ interface IOddzAsset {
     function addAsset(
         bytes32 _name,
         address _address,
-        uint256 _precision
+        uint8 _precision
     ) external returns (uint32 assetId);
 
     /**
